src/requirements.txt
dj-database-url==0.5.0
Django==1.9
django-crispy-forms==1.6.0
django-markdown-deux==1.0.5
django-pagedown==0.1.1
gunicorn==19.8.1
markdown2==2.3.1
Pillow==5.1.0
psycopg2==2.7.5
whitenoise==3.3.1
